crewai/requirements.txt
crewai
crewai-tools
langchain_mistralai
langchain_google_genai
load_dotenv
vllm